+++ tests/testthat/test-to_xml.R
@@ -1,7 +1,7 @@
 test_that("to_xml works", {
   path <- system.file("extdata", "example1.md", package = "tinkr")
   post_list <- to_xml(path)
-  expect_equal(names(post_list)[1], "yaml")
+  expect_equal(names(post_list)[1], "frontmatter")
   expect_equal(names(post_list)[2], "body")
   expect_s3_class(post_list[[2]], "xml_document")
 })
@@ -19,7 +19,7 @@ test_that("#65 to_xml does not throw a warning for no newline", {
 test_that("to_xml works for Rmd", {
   path <- system.file("extdata", "example2.Rmd", package = "tinkr")
   post_list <- to_xml(path)
-  expect_equal(names(post_list)[1], "yaml")
+  expect_equal(names(post_list)[1], "frontmatter")
   expect_equal(names(post_list)[2], "body")
   expect_s3_class(post_list[[2]], "xml_document")
 
@@ -52,8 +52,8 @@ test_that("to_xml will not convert numeric options to character", {
   con <- textConnection(txt)
   code <- xml2::xml_find_first(to_xml(con)$body, "d1:code_block")
   attrs <- xml2::xml_attrs(code)
-  expect_equal(attrs[["fig.width"]], "4.2") 
-  expect_equal(attrs[["fig.height"]], "4.2") 
+  expect_equal(attrs[["fig.width"]], "4.2")
+  expect_equal(attrs[["fig.height"]], "4.2")
   # out.width is the only one that's quoted
   expect_equal(attrs[["out.width"]], shQuote("100%", type = "cmd"))
   expect_equal(attrs[["purl"]], "TRUE")
@@ -91,14 +91,14 @@ test_that("to_xml works with text connection", {
   path <- system.file("extdata", "example2.Rmd", package = "tinkr")
   txt  <- readLines(path)
   con  <- textConnection(txt)
-  expect_equal(to_xml(path)$yaml, to_xml(con)$yaml)
+  expect_equal(to_xml(path)$frontmatter, to_xml(con)$frontmatter)
 
 })
 
 test_that("to_xml works with sourcepos", {
   path <- system.file("extdata", "example1.md", package = "tinkr")
   post_list <- to_xml(path, sourcepos = TRUE)
-  expect_equal(names(post_list)[1], "yaml")
+  expect_equal(names(post_list)[1], "frontmatter")
   expect_equal(names(post_list)[2], "body")
   expect_s3_class(post_list[[2]], "xml_document")
   expect_true(xml2::xml_has_attr(post_list[[2]], "sourcepos"))
@@ -108,7 +108,7 @@ test_that("to_xml works with sourcepos", {
 test_that("to_xml works with sourcepos for Rmd", {
   path <- system.file("extdata", "example2.Rmd", package = "tinkr")
   post_list <- to_xml(path, sourcepos = TRUE)
-  expect_equal(names(post_list)[1], "yaml")
+  expect_equal(names(post_list)[1], "frontmatter")
   expect_equal(names(post_list)[2], "body")
   expect_s3_class(post_list[[2]], "xml_document")
   expect_true(xml2::xml_has_attr(post_list[[2]], "sourcepos"))