+++ NEWS.md
@@ -22,7 +22,13 @@
   (@zkamvar, #111; reviewed: @maelle)
 * Block math will now include the delimiters and the softbreaks for protection
   (issue/review: #113, @maelle; implemented: #111, @zkamvar)
-
+* Frontmatter in TOML and JSON is now preserved too. 
+  The `yaml` field is replaced (softly deprecated)
+  with a field called `frontmatter`. 
+  A new field called `frontmatter_format` ("YAML", "TOML", or "JSON")
+  contains the format of the frontmatter.
+  (issue: #126, @maelle)
+  
 ## NEW IMPORTS
 
 * We now import {rlang} for error handling. Because we already import {purrr},
@@ -41,7 +47,8 @@
 * Documents with no trailing newline will no longer throw a warning (issue: #65;
   fix: #114, @zkamvar)
 * Documents with dollar signs but no math will no longer fail with the
-  `$protect_math()` method (issue: #121, @maelle; fix: #122, @zkamvar).
+  `$protect_math()` method
+  (issue: #121, #124 @maelle; fix: #122, #125 @zkamvar)
 
 ## MISC